# docs/api.yaml
openapi: 3.0.3
info:
  title: CareConnect API
  description: |
    API for managing medication prescriptions and sending notifications to elderly users.
    This system helps users and caretakers manage medication schedules through timely notifications.
  version: 1.0.0
  contact:
    name: API Support
    email: support@mednotify.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://careconnect.lmezencio.dev/api/v1
    description: Production server
...
security:
  - FirebaseJWT: []
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: Register a new user in the system
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      description: |
        Authenticate a user with email and password. Returns a Firebase ID
        token (JWT) that the client must send as `Authorization: Bearer <token>`
        on subsequent authenticated requests. Push notifications use device
        tokens managed via the `/users/me/device-tokens` endpoints.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /prescriptions:
    post:
      tags:
        - Prescriptions
      summary: Create a new prescription
      description: |
        Create a new prescription for a user. This will automatically schedule
        notifications based on the medicament times specified.
      operationId: createPrescription
      security:
        - DemoSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrescriptionRequest'
            examples:
              once_daily:
                summary: Once daily medication for 7 days
                value:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  medic_id: "660e8400-e29b-41d4-a716-446655440001"
                  medicaments:
                    - name: "Aspirin"
                      dosage: "100mg"
                      frequency: "24:00"
                      time: ["08:00"]
                      doses: 7
              twice_daily:
                summary: Twice daily medication for 7 days
                value:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  medic_id: "660e8400-e29b-41d4-a716-446655440001"
                  medicaments:
                    - name: "Lisinopril"
                      dosage: "10mg"
                      frequency: "12:00"
                      time: ["08:00", "20:00"]
                      doses: 14
              multiple_medications:
                summary: Multiple medications with different schedules
                value:
                  user_id: "550e8400-e29b-41d4-a716-446655440000"
                  medic_id: "660e8400-e29b-41d4-a716-446655440001"
                  medicaments:
                    - name: "Aspirin"
                      dosage: "100mg"
                      frequency: "24:00"
                      time: ["08:00"]
                      doses: 30
                    - name: "Lisinopril"
                      dosage: "10mg"
                      frequency: "12:00"
                      time: ["08:00", "20:00"]
                      doses: 60
                    - name: "Metformin"
                      dosage: "500mg"
                      frequency: "08:00"
                      time: ["06:00", "14:00", "22:00"]
                      doses: 21
      responses:
        '201':
          description: Prescription created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Prescription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: User or doctor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags:
        - Prescriptions
      summary: List prescriptions
      description: Retrieve prescriptions with optional filters
      operationId: listPrescriptions
      parameters:
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: string
            format: uuid
        - name: medic_id
          in: query
          description: Filter by doctor ID
          schema:
            type: string
            format: uuid
        - name: active
          in: query
          description: Filter by active status
          schema:
            type: boolean
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Prescription'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    FirebaseJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Firebase JWT token for authentication. The `Authorization` header
        accepts either the standard form `Bearer <token>` (scheme
        case-insensitive) or the bare Firebase ID token returned by
        `/auth/login` — both are accepted so clients can paste the token
        straight from the login response without adding a prefix.
        Non-Bearer schemes (e.g. `Basic`, `Token`) and malformed tokens
        are rejected with `401 invalid authorization format`.
      x-issuer: Firebase

    DemoSecret:
      type: apiKey
      name: Authorization
      in: header
      description: Demo secret for prescription creation
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        name:
          type: string
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john.doe@example.com"
        phone:
          type: string
          description: User's phone number
          example: "+1234567890"
        cpf:
          type: string
          description: |
            Brazilian CPF (Cadastro de Pessoa Física), digits-only.
            May be empty if the user was registered without one.
          example: "52998224725"
        created_at:
          type: string
          format: date-time
          description: User creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - name
        - email
        - phone
        - created_at
        - updated_at
...
    Prescription:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique prescription identifier
        user_id:
          type: string
          format: uuid
          description: ID of the user this prescription belongs to
        medic_id:
          type: string
          format: uuid
          description: ID of the doctor who created this prescription
        medicaments:
          type: array
          description: List of medicaments in this prescription
          items:
            $ref: '#/components/schemas/Medicament'
        active:
          type: boolean
          description: Whether this prescription is active
          example: true
        created_at:
          type: string
          format: date-time
          description: Prescription creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - user_id
        - medic_id
        - medicaments
        - active
        - created_at
        - updated_at
    CreatePrescriptionRequest:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: ID of the user this prescription is for
          example: "550e8400-e29b-41d4-a716-446655440000"
        medic_id:
          type: string
          format: uuid
          description: ID of the doctor creating this prescription
          example: "660e8400-e29b-41d4-a716-446655440001"
        medicaments:
          type: array
          description: List of medicaments in this prescription
          minItems: 1
          items:
            $ref: '#/components/schemas/Medicament'
      required:
        - user_id
        - medic_id
        - medicaments
...
    Medicament:
      type: object
      properties:
        name:
          type: string
          description: Name of the medication
          example: "Aspirin"
        dosage:
          type: string
          description: Dosage amount and unit
          example: "100mg"
        frequency:
          type: string
          description: |
            Duration between doses in HH:MM format.
            - "24:00" = once daily
            - "12:00" = twice daily (every 12 hours)
            - "08:00" = three times daily (every 8 hours)
            - "06:00" = four times daily (every 6 hours)
          pattern: '^([0-1][0-9]|2[0-4]):[0-5][0-9]$'
          example: "24:00"
        time:
          type: array
          description: |
            Specific times to take the medication in HH:MM format (24-hour).
            The number of times must match the frequency:
            - frequency "24:00" requires 1 time
            - frequency "12:00" requires 2 times
            - frequency "08:00" requires 3 times
            - frequency "06:00" requires 4 times
          minItems: 1
          items:
            type: string
            pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
          example: ["08:00"]
        doses:
          type: integer
          minimum: 1
          description: |
            Total number of doses for this medicament.
            This determines when the prescription ends.
            Examples:
            - 7 doses with frequency "24:00" = 7 days (once daily for a week)
            - 14 doses with frequency "12:00" = 7 days (twice daily for a week)
            - 21 doses with frequency "08:00" = 7 days (three times daily for a week)
          example: 7
      required:
        - name
        - dosage
        - frequency
        - time
        - doses

    # Error schemas
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request"
        details:
          type: string
          description: Additional error details
          example: "Field 'name' is required"
      required:
        - error

    # Auth schemas
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
          description: User's full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User's email address
          example: "john.doe@example.com"
        phone:
          type: string
          description: User's phone number
          example: "+1234567890"
        cpf:
          type: string
          description: |
            Brazilian CPF (Cadastro de Pessoa Física) — 11 digits.
            Optional, but when present is validated using the official
            modulo-11 checksum and normalized to digits-only.
            Accepts formatted input such as "529.982.247-25".
          example: "529.982.247-25"
        password:
          type: string
          format: password
          description: User's password
          example: "securepassword123"
        role:
          type: string
          enum: [ELDERLY, CAREGIVER]
          description: |
            User role. Optional, defaults to "ELDERLY" if omitted or invalid.
          example: "ELDERLY"
      required:
        - name
        - email
        - phone
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: "john.doe@example.com"
        password:
          type: string
          format: password
          description: User's password
          example: "securepassword123"
      required:
        - email
        - password
    AuthResponse:
      type: object
      description: |
        Authentication response returned by `/auth/login` and
        `/auth/login/doctor`. The `token` is a Firebase ID token (JWT) that
        must be sent as `Authorization: Bearer <token>` on subsequent
        authenticated requests. Push notifications use device tokens managed
        via the `/users/me/device-tokens` endpoints.
      properties:
        token:
          type: string
          description: Firebase ID token (JWT) to be used as bearer token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: '#/components/schemas/User'
      required:
        - token
        - user

    # LGPD / GDPR data-export schemas
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request"
            details: "Field 'name' is required"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            details: "An unexpected error occurred"
    ServiceUnavailable:
      description: Service unavailable - server is temporarily unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Service unavailable"
            details: "The server is temporarily unavailable. Please try again later."